Add a main class for manage classes and improve structure
