+++ src/main/java/dev/alejandrosahonero/gui/options/manage/ModAlumView.java
@@ -1,20 +1,18 @@
 package dev.alejandrosahonero.gui.options.manage;
 
 import dev.alejandrosahonero.gui.MainView;
+import dev.alejandrosahonero.logic.options.manage.ModModController;
+import dev.alejandrosahonero.logic.options.manage.PhotoButtonController;
+import dev.alejandrosahonero.logic.options.manage.matricular.MatricularButtonController;
+import dev.alejandrosahonero.logic.options.manage.matricular.txtController;
 
 import javax.swing.*;
+import java.awt.*;
+import java.util.List;
 
-public class ModAlumView extends MainView {
-    private JLabel background;
-
+public class ModAlumView extends MainManageView {
     public ModAlumView(){
         super("Modificar Alumno");
-        setDefaultCloseOperation(JFrame.HIDE_ON_CLOSE);
-
-        initComponents();
-    }
-    private void initComponents(){
-        background = new JLabel(getBackgroundImage());
-        add(background);
+        super.doneButton("MODIFICAR");
     }
 }
