+++ src/main/java/dev/alejandrosahonero/gui/options/manage/ModAsigView.java
@@ -4,17 +4,9 @@ import dev.alejandrosahonero.gui.MainView;
 
 import javax.swing.*;
 
-public class ModAsigView extends MainView {
-    private JLabel background;
-
+public class ModAsigView extends MainManageView {
     public ModAsigView(){
         super("Modificar Asignatura");
-        setDefaultCloseOperation(JFrame.HIDE_ON_CLOSE);
-
-        initComponents();
-    }
-    private void initComponents(){
-        background = new JLabel(getBackgroundImage());
-        add(background);
+        super.doneButton("MODIFICAR");
     }
 }
