+++ src/main/java/dev/alejandrosahonero/gui/options/manage/ModModView.java
@@ -5,22 +5,9 @@ import dev.alejandrosahonero.gui.MainView;
 import javax.swing.*;
 import java.awt.*;
 
-public class ModModView extends MainView {
-    JLabel background;
-
-    public ModModView() {
-        super("Moduficar Módulos");
-        setDefaultCloseOperation(JFrame.HIDE_ON_CLOSE);
-
-        initComponents();
-    }
-
-    /**
-     * Método que inicializa todos los demás métodos para aligerar la carga al constructor
-     */
-    private void initComponents() {
-        background = new JLabel(getBackgroundImage());
-
-        add(background, BorderLayout.CENTER);
+public class ModModView extends MainManageView {
+    public ModModView(){
+        super("Modificar Módulo");
+        super.doneButton("MODIFICAR");
     }
 }
