+++ src/main/java/dev/alejandrosahonero/gui/options/manage/NewAsigView.java
@@ -7,20 +7,9 @@ import dev.alejandrosahonero.logic.options.OptionsController;
 import javax.swing.*;
 import java.awt.*;
 
-public class NewAsigView extends MainView {
-    private JLabel background;
-
+public class NewAsigView extends MainManageView {
     public NewAsigView(){
         super("Añadir Asignatura");
-        setDefaultCloseOperation(JFrame.HIDE_ON_CLOSE);
-
-        initComponents();
-    }
-
-    private void initComponents(){
-        background = new JLabel(getBackgroundImage());
-
-
-        add(background, BorderLayout.CENTER);
+        super.doneButton("AÑADIR");
     }
 }
