+++ src/main/java/dev/alejandrosahonero/gui/options/manage/NewProView.java
@@ -6,20 +6,9 @@ import dev.alejandrosahonero.logic.options.OptionsController;
 import javax.swing.*;
 import java.awt.*;
 
-public class NewProView extends MainView {
-    private JLabel background;
-
+public class NewProView extends MainManageView {
     public NewProView(){
         super("Añadir Profesor");
-        setDefaultCloseOperation(JFrame.HIDE_ON_CLOSE);
-
-        initComponents();
-    }
-
-    private void initComponents(){
-        background = new JLabel(getBackgroundImage());
-
-
-        add(background, BorderLayout.CENTER);
+        super.doneButton("AÑADIR");
     }
 }
